Refactor research filtering tool and update namespace
Refactored `research_filtering` to use the `TreeOfLife_toolbox` namespace, replacing `DD_tools`. Enhanced documentation across classes, methods, and the README to improve clarity and maintainability. Adjusted configurations and paths to align with new processing workflows.
