--- src/DD_tools/research_filtering/README.md
@@ -1,7 +1,0 @@
-tool to fix the column name mistake in `research_filtering`
-
-Additional config fields:
-
-* `occurrences_path` - path to uuid table with duplicated uuids
-* `data_path` - path to data folder
-* `basis_of_record` - basis of records that are going to be filtered
