+++ src/TreeOfLife_toolbox/__init__.py
@@ -1,0 +1,1 @@
+from TreeOfLife_toolbox import research_filtering
