+++ src/TreeOfLife_toolbox/research_filtering/__init__.py
@@ -1,0 +1,5 @@
+from .classes import (
+    ResearchFilteringFilter,
+    ResearchFilteringRunner,
+    ResearchFilteringScheduleCreation,
+)
